pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "mr-prog-discord-bot"
authors = [
    {name = "wchill"}
]
description = "Mr. Prog Discord Bot"
requires-python = ">=3.8"
version = "1.0"
dependencies = [
    "aio-pika",
    "asyncio-mqtt",
    "discord.py",
    "jsonpickle",
    "mmbn-gamedata",
    "mr-prog-utils",
    "psutil",
    "py-cpuinfo",
    "pillow"
]

[tool.setuptools]
platforms = ["linux"]

[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
py_version = 38
